requirements.txt
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
dj-database-url==0.4.2
Django==1.11
djangorestframework==3.8.2
dropbox==8.7.1
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
olefile==0.44
Pillow==4.1.1
protobuf==3.3.0
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pytz==2017.2
requests==2.20.1
rsa==3.4.2
six==1.10.0
urllib3==1.21.1
virtualenv==15.1.0
Werkzeug==0.14.1
whitenoise==3.3.0
